+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "healthbot-container",
-  "version": "1.3.2",
+  "version": "1.3.3",
   "description": "Health Backbone webchat client",
   "main": "server/server.js",
   "dependencies": {
